+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 0.11.4
 
-* stax pull - changes --stay-on-head to --no-switch-back (-n) to avoid clashing with global help flag
 * stax any - adds `--log` global flag to automatically execute `stax log` after finishing the requested command
-* stax pull - adds --stay-on-head flag to allow staying on the head/default branch after pulling
+* stax pull - adds -n/--no-switch-back flag to allow staying on the head/default branch after pulling
 
 0.11.3