+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.11.4
 
 * stax commit - adds -c/--come-back flag to move back to the branch on which user was before running commit
+* stax extras doctor - adds -j/--json flag to output in json format
 * stax extras nuke - new command that resets working directory and index to HEAD and cleans all untracked files
 * stax * - adds `--log` global flag to automatically execute `stax log` after finishing the requested command
 * stax pull - adds -n/--no-switch-back flag to allow staying on the head/default branch after pulling